requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
djangorestframework==3.14.0
flake8==6.1.0
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
magic-filter==1.0.11
mccabe==0.7.0
multidict==6.0.4
pycodestyle==2.11.0
pyflakes==3.1.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
yarl==1.9.2
